refactor: extract v2 eval context
